Moved avatar to its own component
