+++ screenit/src/components/Avatar/style.css
@@ -1,0 +1,54 @@
+/* ————————————————————————————————————————————————
+Table Of Contents
+
+1. Main Code 
+2. Breakpoints 
+
+———————————————————————————————————————————————— */
+
+
+
+/* ————————————————————————————————————————————————
+1. Main Code
+———————————————————————————————————————————————— */
+
+
+.avatar {
+    background: var(--primary);
+    color: white;
+    text-transform: uppercase;
+    height: 64px;
+    width: 64px;
+    border-radius: 50px;
+    border: 1px solid #7AAFFF;
+    box-sizing: border-box;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-weight: 500;
+    box-shadow: 0px 4px 8px -1px rgba(97, 160, 255, 0.2), 0px 6px 10px rgba(97, 160, 255, 0.25), 0px 2px 0px rgba(97, 160, 255, 0.5);
+}
+
+.avatar--lg {
+    font-size: 24px;
+}
+
+.avatar--sm {
+    font-size: 16px;
+}
+
+
+/* ————————————————————————————————————————————————
+2. Breakpoints 
+———————————————————————————————————————————————— */
+
+
+@media screen and (max-width: 600px) {
+
+    .avatar {
+        height: 45px;
+        width: 45px;
+    }
+
+}
+
