+++ screenit/src/components/Form/Form.js
@@ -78,7 +78,7 @@ const Form = () => {
         axios.get(omdbUrl)
             .then((res) => {
 
-                // console.log(res.data)
+                console.log(res.data)
 
                 // Stores IMDB ID
                 let imdbMovieID = JSON.stringify(res.data.imdbID);
