+++ screenit/src/components/ResultRow/ResultRow.js
@@ -4,6 +4,7 @@ import RatingStars from '../RatingStars/RatingStars';
 import { useSeenItContext } from '../../utils/SeenItContext';
 import { FaPlus } from "react-icons/fa";
 import ResultTools from "../ResultTools/ResultTools";
+import Avatar from "../Avatar/Avatar";
 
 
 const ResultRow = () => {
@@ -29,33 +30,6 @@ const ResultRow = () => {
             <div className="results__overflow">
                 {
                     resultsArray.map((result, index) => {
-
-
-                        // START Avatar —————————————————————————————————|
-
-                        // Take the movie's name and creates a new array, splitting on the spaces and special characters
-                        let avatarArray = result.movieName.trim().split(/[^A-Z0-9]/ig).slice(0, 3);
-
-                        // Stores the final 1-3 characters from the array
-                        const avatarCharacters = [];
-
-                        // Loops through array to grab the first character of each word in the array
-                        avatarArray.forEach(e => {
-                            avatarCharacters.push(e.charAt(0));
-                        });
-
-                        // Variable that holds a "lg" or "sm" value to decide the right class
-                        let avatarLength = "";
-
-                        // Decides which size based on avatar length 
-                        if (avatarCharacters.length > 2) {
-                            avatarLength = "sm";
-                        } else {
-                            avatarLength = "lg";
-                        }
-
-                        // END Avatar —————————————————————————————————|
-
 
 
                         // START Result Slider —————————————————————————————————|
@@ -123,11 +97,11 @@ const ResultRow = () => {
                                     onClick={() => slideItemHandle(index)} >
 
                                     {/* Avatar */}
-                                    < span className={`result-row__avatar ${avatarLength === 'sm' ? 'result-row__avatar--sm' : 'result-row__avatar--lg'}`}>{avatarCharacters}</span>
+                                    <Avatar movieName={result.movieName} />
 
                                     {/* Name and Category */}
                                     <div className="result-row__txt">
-                                        <h2 className="result-row__movie-name">{result.movieName.trim('')}</h2>
+                                        <h2 className="result-row__movie-name">{result.movieName}</h2>
                                         <h3 className="result-row__category">{result.category}</h3>
                                     </div>
